dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    make dev-requirements.txt
#
build==1.1.1
    # via pip-tools
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via pip-tools
distlib==0.3.8
    # via virtualenv
exceptiongroup==1.2.0
    # via pytest
filelock==3.13.1
    # via virtualenv
identify==2.5.35
    # via pre-commit
iniconfig==2.0.0
    # via pytest
nodeenv==1.8.0
    # via
    #   pre-commit
    #   pyright
packaging==23.2
    # via
    #   build
    #   pytest
pip-tools==7.4.0
    # via -r dev-requirements.in
platformdirs==4.2.0
    # via virtualenv
pluggy==1.4.0
    # via pytest
pre-commit==3.6.2
    # via -r dev-requirements.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pyright==1.1.365
    # via -r dev-requirements.in
pytest==8.0.2
    # via -r dev-requirements.in
pyyaml==6.0.1
    # via pre-commit
ruff==0.4.7
    # via -r dev-requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
virtualenv==20.25.1
    # via pre-commit
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
